Moved shared TableHealth logic to its own class
